fix: refactor new packages added
