+++ app/src/main/java/com/github/se/assocify/ui/screens/treasury/TreasuryScreen.kt
@@ -56,8 +56,8 @@ import com.github.se.assocify.navigation.Destination
 import com.github.se.assocify.navigation.MAIN_TABS_LIST
 import com.github.se.assocify.navigation.NavigationActions
 import com.github.se.assocify.ui.composables.MainNavigationBar
-import com.github.se.assocify.ui.screens.treasury.accounting.Balance
-import com.github.se.assocify.ui.screens.treasury.accounting.Budget
+import com.github.se.assocify.ui.screens.treasury.accounting.balance.Balance
+import com.github.se.assocify.ui.screens.treasury.accounting.budget.Budget
 import com.github.se.assocify.ui.util.DateUtil
 import com.github.se.assocify.ui.util.PriceUtil
 import kotlinx.coroutines.launch
