+++ app/src/main/java/com/github/se/assocify/ui/screens/treasury/accounting/balance/BalanceScreen.kt
@@ -1,6 +1,7 @@
-package com.github.se.assocify.ui.screens.treasury.accounting
+package com.github.se.assocify.ui.screens.treasury.accounting.balance
 
 import androidx.compose.runtime.Composable
+import com.github.se.assocify.ui.screens.treasury.accounting.Accounting
 
 /** The accounting screen displaying the balance screen of the association */
 @Composable
